remove obsolete test class and consolidate tests
