+++ OutOfSchool/OutOfSchool.WebApi.Tests/Services/ExternalExportServiceTests.cs
@@ -14,7 +14,6 @@ using OutOfSchool.BusinessLogic.Util.Mapping;
 using OutOfSchool.Common.Enums.CompetitiveEvent;
 using OutOfSchool.Services.Models;
 using OutOfSchool.Services.Models.CompetitiveEvents;
-using OutOfSchool.Services.Models.SubordinationStructure;
 using OutOfSchool.Services.Repository.Api;
 using OutOfSchool.Services.Repository.Base.Api;
 using OutOfSchool.Tests.Common;
@@ -30,7 +29,6 @@ public class ExternalExportServiceTests
     private Mock<IApplicationRepository> mockApplicationRepository;
     private Mock<IAverageRatingService> mockAverageRatingService;
     private Mock<IEntityRepositorySoftDeleted<long, Direction>> mockDirectionRepository;
-    private Mock<ISensitiveEntityRepositorySoftDeleted<Institution>> mockInstitutionRepository;
     private Mock<ISensitiveEntityRepositorySoftDeleted<CompetitiveEvent>> mockCompetitiveEventRepository;
     private Mock<IEntityRepositorySoftDeleted<long, SubDirection>> mockSubDirectionRepository;
     private IMapper mockMapper;
@@ -44,7 +42,6 @@ public class ExternalExportServiceTests
         mockApplicationRepository = new Mock<IApplicationRepository>();
         mockAverageRatingService = new Mock<IAverageRatingService>();
         mockDirectionRepository = new Mock<IEntityRepositorySoftDeleted<long, Direction>>();
-        mockInstitutionRepository = new Mock<ISensitiveEntityRepositorySoftDeleted<Institution>>();
         mockCompetitiveEventRepository = new Mock<ISensitiveEntityRepositorySoftDeleted<CompetitiveEvent>>();
         mockSubDirectionRepository = new Mock<IEntityRepositorySoftDeleted<long, SubDirection>>();
         mockMapper = TestHelper.CreateMapperInstanceOfProfileTypes<CommonProfile, ExternalExportMappingProfile>();
@@ -320,22 +317,6 @@ public class ExternalExportServiceTests
         Assert.CatchAsync<Exception>(() => externalExportService.GetDirections(updatedAfter, new OffsetFilter()));
     }
 
-    /// <summary>
-    /// This is the only sub direction logic, that can be tested on mocks
-    /// </summary>
-    [Test]
-    public void GetSubDirections_ExceptionInGetSubDirections_ReturnsEmptySearchResult()
-    {
-        // Arrange
-        var updatedAfter = DateTime.UtcNow;
-        var offsetFilter = new OffsetFilter { Size = 10 };
-        mockSubDirectionRepository.Setup(repo => repo.Get(offsetFilter.From, offsetFilter.Size, It.IsAny<string>(), It.IsAny<Func<IQueryable<SubDirection>, IQueryable<SubDirection>>>(), It.IsAny<Expression<Func<SubDirection,bool>>>(), null, false))
-            .Throws(new Exception("Simulated exception"));
-
-        // Act & Assert
-        Assert.CatchAsync<Exception>(() => externalExportService.GetSubDirections(updatedAfter, new OffsetFilter()));
-    }
-
     [Test]
     public void Constructor_NullProviderRepository_ThrowsArgumentNullException()
     {
@@ -382,6 +363,71 @@ public class ExternalExportServiceTests
         // Arrange, Act, Assert
         Assert.Throws<ArgumentNullException>(() =>
             new ExternalExportService(Mock.Of<IProviderRepository>(), Mock.Of<IWorkshopRepository>(), Mock.Of<IApplicationRepository>(), null, null, null, null, Mock.Of<IMapper>(), Mock.Of<ILogger<ExternalExportService>>()));
+    }
+    
+    [Test]
+    public async Task GetSubDirections_ReturnsEmptySearchResult()
+    {
+        // Arrange
+        var updatedAfter = DateTime.UtcNow;
+        var offsetFilter = new OffsetFilter { Size = 10 };
+        var fakeSubDirections = new List<SubDirection>();
+
+        mockSubDirectionRepository
+            .Setup(x => x.Get(offsetFilter.From, offsetFilter.Size, It.IsAny<string>(), It.IsAny<Func<IQueryable<SubDirection>, IQueryable<SubDirection>>>(), It.IsAny<Expression<Func<SubDirection, bool>>>(), null, false))
+            .Returns(fakeSubDirections.AsTestAsyncEnumerableQuery());
+
+        // Act
+        var result = await externalExportService.GetSubDirections(updatedAfter, offsetFilter);
+
+        // Assert
+        Assert.IsNotNull(result);
+        Assert.AreEqual(0, result.TotalAmount);
+        Assert.AreEqual(0, result.Entities.Count);
+    }
+
+    [Test]
+    public async Task GetSubDirections_ReturnsSearchResultData()
+    {
+        // Arrange
+        var updatedAfter = DateTime.UtcNow;
+        var offsetFilter = new OffsetFilter { Size = 10 };
+
+        var fakeSubDirections = new List<SubDirection>
+        {
+            new() { Id = 1, Title = "SubDirection 1" },
+            new() { Id = 2, Title = "SubDirection 2" },
+            new() { Id = 3, Title = "SubDirection 3" },
+            new() { Id = 4, Title = "SubDirection 4" }
+        };
+
+        mockSubDirectionRepository
+            .Setup(x => x.Get(offsetFilter.From, offsetFilter.Size, It.IsAny<string>(), It.IsAny<Func<IQueryable<SubDirection>, IQueryable<SubDirection>>>(), It.IsAny<Expression<Func<SubDirection, bool>>>(), null, false))
+            .Returns(fakeSubDirections.AsTestAsyncEnumerableQuery());
+
+        mockSubDirectionRepository.Setup(x => x.Count(It.IsAny<Expression<Func<SubDirection, bool>>>())).ReturnsAsync(fakeSubDirections.Count);
+
+        // Act
+        var result = await externalExportService.GetSubDirections(default, offsetFilter);
+
+        // Assert
+        Assert.IsNotNull(result);
+        Assert.AreEqual(fakeSubDirections.Count, result.TotalAmount);
+        Assert.AreEqual(fakeSubDirections.Count, result.Entities.Count);
+        mockSubDirectionRepository.Verify(x => x.Count(It.IsAny<Expression<Func<SubDirection, bool>>>()), Times.Once);
+    }
+
+    [Test]
+    public void GetSubDirections_ExceptionInGetSubDirections_ReturnsEmptySearchResult()
+    {
+        // Arrange
+        var updatedAfter = DateTime.UtcNow;
+        var offsetFilter = new OffsetFilter { Size = 10 };
+        mockSubDirectionRepository.Setup(repo => repo.Get(offsetFilter.From, offsetFilter.Size, It.IsAny<string>(), It.IsAny<Func<IQueryable<SubDirection>, IQueryable<SubDirection>>>(), It.IsAny<Expression<Func<SubDirection, bool>>>(), null, false))
+            .Throws(new Exception("Simulated exception"));
+
+        // Act & Assert
+        Assert.CatchAsync<Exception>(() => externalExportService.GetSubDirections(updatedAfter, new OffsetFilter()));
     }
 
     private List<CompetitiveEvent> CompetitiveEvents()
